feat : api마무리, 디렉토리 수정, getWeekRange구현
