+++ src/layouts/Dropdown.tsx
@@ -1,0 +1,110 @@
+import React from "react";
+import Button from "@mui/material/Button";
+import ClickAwayListener from "@mui/material/ClickAwayListener";
+import Grow from "@mui/material/Grow";
+import Paper from "@mui/material/Paper";
+import Popper from "@mui/material/Popper";
+import MenuItem from "@mui/material/MenuItem";
+import MenuList from "@mui/material/MenuList";
+import { addDays } from "date-fns";
+
+const Dropdown = () => {
+  const [open, setOpen] = React.useState<boolean>(false);
+  const anchorRef = React.useRef<HTMLButtonElement>(null);
+
+  //주간 산출하는 함수
+  //반복문 => 돌리고 endDate이 4월 20일 기준 7일미만이면 반복 탈출
+  const getWeekRange = () : string[][] => {
+    const weeks = [];
+    const lastDate: Date = new Date("2022-04-20");
+    let startDate: Date = new Date("2022-02-01");
+    while (startDate <= lastDate) {
+      let endDate: Date = addDays(startDate, 6);
+      weeks.push([startDate.toString(), endDate.toString()]);
+      startDate = addDays(startDate, 7);
+    }
+    return weeks;
+  };
+  console.log(getWeekRange());
+
+  const handleToggle = () => {
+    setOpen((prevOpen) => !prevOpen);
+  };
+
+  const handleClose = (event: any) => {
+    if (anchorRef.current && anchorRef.current.contains(event.target)) {
+      return;
+    }
+
+    setOpen(false);
+  };
+
+  function handleListKeyDown(event: React.KeyboardEvent) {
+    if (event.key === "Tab") {
+      event.preventDefault();
+      setOpen(false);
+    } else if (event.key === "Escape") {
+      setOpen(false);
+    }
+  }
+
+  // return focus to the button when we transitioned from !open -> open
+  const prevOpen = React.useRef(open);
+  React.useEffect(() => {
+    if (prevOpen.current === true && open === false) {
+      anchorRef.current?.focus();
+    }
+
+    prevOpen.current = open;
+  }, [open]);
+
+  return (
+    <div>
+      <Button
+        ref={anchorRef}
+        id="composition-button"
+        aria-controls={open ? "composition-menu" : undefined}
+        aria-expanded={open ? "true" : undefined}
+        aria-haspopup="true"
+        onClick={handleToggle}
+      >
+        날짜
+      </Button>
+      <Popper
+        open={open}
+        anchorEl={anchorRef.current}
+        role={undefined}
+        placement="bottom-start"
+        transition
+        disablePortal
+      >
+        {({ TransitionProps, placement }) => (
+          <Grow
+            {...TransitionProps}
+            style={{
+              transformOrigin:
+                placement === "bottom-start" ? "left top" : "left bottom",
+            }}
+          >
+            <Paper>
+              <ClickAwayListener onClickAway={handleClose}>
+                <MenuList
+                  autoFocusItem={open}
+                  id="composition-menu"
+                  aria-labelledby="composition-button"
+                  onKeyDown={handleListKeyDown}
+                >
+                  <MenuItem onClick={handleClose}>2월 1일 ~ 2월7일</MenuItem>
+                  <MenuItem onClick={handleClose}>2월 8일 ~ 2월14일</MenuItem>
+                  <MenuItem onClick={handleClose}>2월 15일 ~ 2월21일</MenuItem>
+                </MenuList>
+              </ClickAwayListener>
+            </Paper>
+          </Grow>
+        )}
+      </Popper>
+    </div>
+  );
+};
+
+export default Dropdown;
