+++ src/pages/Dashboard.tsx
@@ -1,8 +1,7 @@
-import React from "react";
 import { Box, Toolbar } from "@mui/material";
-import TotalAdStatus from "../components/TotalAdStatus";
-import Dropdown from "../components/Dropdown";
-//import ChannelStatus from "../components/ChannelStatus";
+import TotalAdStatus from "../components/dashboard/totalAdStatus/TotalAdStatus";
+import Dropdown from "../layouts/Dropdown";
+import ChannelStatus from "../components/dashboard/channelStatus/ChannelStatus";
 
 const Dashboard = () => {
   return (
@@ -15,12 +14,9 @@ const Dashboard = () => {
         width: `calc(100vw - 240px)`,
       }}
     >
-      <Toolbar>대시 보드</Toolbar>
-      {/* 드롭다운 */}  
-      <Dropdown />
-      
+      <Toolbar>대시 보드 <Dropdown /></Toolbar>     
       <TotalAdStatus />
-      {/* <ChannelStatus /> */}
+      <ChannelStatus />
     </Box>
   );
 };
